UI Overhaul
changed the UI, Buy Menu and various other changes.

+++ Assets/Scripts/DepositButton.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class DepositButton : MonoBehaviour 
 {
-	public double transferAmount;
-	public double bankDepositFee;
+	public float transferAmount;
+	public float bankDepositFee;
 
 	public void depositCash()
 	{
